refactor(catalog): extract component queries
